Device List - datagrid
