+++ HMB_Utility/Converters/IsConnectedToFontSizeConverter.cs
