+++ HMB_Utility/Converters/isConnectedToFontWeightConverter.cs
