+++ HMB_Utility/FoundModel/FoundSignal.cs
@@ -1,24 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using Hbm.Api.Common;
-using Hbm.Api.Common.Messaging;
-using Hbm.Api.Common.Exceptions;
-using Hbm.Api.Common.Entities;
-using Hbm.Api.Common.Entities.Problems;
+﻿using Hbm.Api.Common.Entities;
 using Hbm.Api.Common.Entities.Connectors;
 using Hbm.Api.Common.Entities.Channels;
 using Hbm.Api.Common.Entities.Signals;
-using Hbm.Api.Common.Entities.Filters;
-using Hbm.Api.Common.Entities.ConnectionInfos;
-using Hbm.Api.Common.Enums;
-using Hbm.Api.Scan;
-using Hbm.Api.Pmx;
-using Hbm.Api.QuantumX;
-using Hbm.Api.Mgc;
 
 namespace HMB_Utility
 {
