+++ HMB_Utility/HMB_Utility.csproj
@@ -110,28 +110,28 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="AppSettings.cs" />
-    <Compile Include="DAQ.cs" />
-    <Compile Include="DeviceInfoConverter.cs" />
-    <Compile Include="Family.cs" />
-    <Compile Include="Filter.cs" />
-    <Compile Include="FoundDevice.cs" />
-    <Compile Include="FoundSignal.cs" />
-    <Compile Include="IpAddress.cs" />
-    <Compile Include="IsConnectedToFontSizeConverter.cs" />
-    <Compile Include="isConnectedToFontWeightConverter.cs" />
-    <Compile Include="Logger.cs" />
-    <Compile Include="MainWindowViewModel.cs" />
-    <Compile Include="HbmMessages.cs" />
-    <Compile Include="Protocol.cs" />
-    <Compile Include="ProtocolEventArg.cs" />
-    <Compile Include="ProtocolMessage.cs" />
-    <Compile Include="ProtocolSaveMethod.cs" />
-    <Compile Include="SignalType.cs" />
-    <Compile Include="UserCommand.cs" />
-    <Compile Include="UserCommandAsync.cs" />
-    <Compile Include="UserCommandAsyncVoid.cs" />
-    <Compile Include="ViewModelBase.cs" />
+    <Compile Include="AppSettings\AppSettings.cs" />
+    <Compile Include="HbmApi\DAQ.cs" />
+    <Compile Include="Converters\DeviceInfoConverter.cs" />
+    <Compile Include="ConnectingViaIp\Family.cs" />
+    <Compile Include="SignalFiltering\Filter.cs" />
+    <Compile Include="FoundModel\FoundDevice.cs" />
+    <Compile Include="FoundModel\FoundSignal.cs" />
+    <Compile Include="ConnectingViaIp\IpAddress.cs" />
+    <Compile Include="Converters\IsConnectedToFontSizeConverter.cs" />
+    <Compile Include="Converters\isConnectedToFontWeightConverter.cs" />
+    <Compile Include="Protocol\Logger.cs" />
+    <Compile Include="ViewModel\MainWindowViewModel.cs" />
+    <Compile Include="Protocol\HbmMessages.cs" />
+    <Compile Include="Protocol\Protocol.cs" />
+    <Compile Include="Protocol\ProtocolEventArg.cs" />
+    <Compile Include="Protocol\ProtocolMessage.cs" />
+    <Compile Include="Protocol\ProtocolSaveMethod.cs" />
+    <Compile Include="SignalFiltering\SignalType.cs" />
+    <Compile Include="Commands\UserCommand.cs" />
+    <Compile Include="Commands\UserCommandAsync.cs" />
+    <Compile Include="Commands\UserCommandAsyncVoid.cs" />
+    <Compile Include="ViewModel\ViewModelBase.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -140,16 +140,16 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="DataToDB.cs" />
-    <Compile Include="HBMFamily.cs" />
-    <Compile Include="HbmSession.cs" />
+    <Compile Include="SavingDataToDB\DataToDB.cs" />
+    <Compile Include="ConnectingViaIp\HBMFamily.cs" />
+    <Compile Include="HbmApi\HbmSession.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Measuring.cs" />
+    <Compile Include="HbmApi\Measuring.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -174,7 +174,7 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <None Include="App.config" />
+    <None Include="AppSettings\App.config" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DB\DB.csproj">
@@ -182,6 +182,7 @@
       <Name>DB</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
