+++ HMB_Utility/HbmApi/HbmSession.cs
@@ -1,24 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using System.Timers;
-using System.Threading;
-using System.Configuration;
-
 using Hbm.Api.Common;
-using Hbm.Api.Common.Messaging;
-using Hbm.Api.Common.Exceptions;
 using Hbm.Api.Common.Entities;
 using Hbm.Api.Common.Entities.Problems;
-using Hbm.Api.Common.Entities.Connectors;
-using Hbm.Api.Common.Entities.Channels;
-using Hbm.Api.Common.Entities.Signals;
-using Hbm.Api.Common.Entities.Filters;
 using Hbm.Api.Common.Entities.ConnectionInfos;
-using Hbm.Api.Common.Enums;
-using Hbm.Api.Scan;
 using Hbm.Api.Pmx;
 using Hbm.Api.QuantumX;
 using Hbm.Api.Mgc;
