+++ HMB_Utility/Protocol/HbmMessages.cs
@@ -1,9 +1,5 @@
 ﻿using Hbm.Api.Common.Messaging;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace HMB_Utility
 {
