+++ HMB_Utility/Protocol/Logger.cs
@@ -1,13 +1,7 @@
-﻿using Hbm.Api.Common.Messaging;
-using Hbm.Api.Logging;
+﻿using Hbm.Api.Logging;
 using Hbm.Api.Logging.Enums;
 using Hbm.Api.Logging.Logger;
 using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace HMB_Utility
 {
