+++ HMB_Utility/Protocol/Protocol.cs
@@ -1,13 +1,5 @@
-﻿using Hbm.Api.Common.Messaging;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Collections.ObjectModel;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Data;
 
 namespace HMB_Utility
