+++ HMB_Utility/Protocol/ProtocolEventArg.cs
@@ -1,10 +1,6 @@
 ﻿using Hbm.Api.Common.Entities.Problems;
-using Hbm.Api.Common.Messaging;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace HMB_Utility
 {
