+++ HMB_Utility/SignalFiltering/SignalType.cs
@@ -1,11 +1,4 @@
-﻿using Hbm.Api.Common.Entities.Signals;
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿
 namespace HMB_Utility
 {
     public class SignalType
