+++ HMB_Utility/ViewModel/MainWindowViewModel.cs
@@ -1,31 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using System.ComponentModel;
-
-using System.Runtime.CompilerServices;
 using System.Collections.ObjectModel;
-
-using Hbm.Api.Common;
-using Hbm.Api.Common.Messaging;
-using Hbm.Api.Common.Exceptions;
 using Hbm.Api.Common.Entities;
-using Hbm.Api.Common.Entities.Problems;
-using Hbm.Api.Common.Entities.Connectors;
-using Hbm.Api.Common.Entities.Channels;
-using Hbm.Api.Common.Entities.Signals;
-using Hbm.Api.Common.Entities.Filters;
-using Hbm.Api.Common.Entities.ConnectionInfos;
-using Hbm.Api.Common.Enums;
-using Hbm.Api.Scan;
-using Hbm.Api.Pmx;
-using Hbm.Api.QuantumX;
-using Hbm.Api.Mgc;
-
-using DB;
-using System.Windows.Data;
 
 namespace HMB_Utility
 {
